extract table, types, and buttons
