+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/ObjectsPage/objectsPageTypes.ts
@@ -1,0 +1,6 @@
+import {KnownBaseObjectClassType} from '../wfReactInterface/wfDataModelHooksInterface';
+
+export type WFHighLevelObjectVersionFilter = {
+  objectName?: string | null;
+  baseObjectClass?: KnownBaseObjectClassType | null;
+};
