+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/ScorersPage/AnnotationsTab.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import {FilterableObjectVersionsTable} from '../ObjectsPage/ObjectVersionsPage';
+import {FilterableObjectVersionsTable} from '../ObjectsPage/ObjectVersionsTable';
 
 export const AnnotationsTab: React.FC<{
   entity: string;
