+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/ScorersPage/LLMJudgesTab.tsx
@@ -2,7 +2,7 @@
 
 import React from 'react';
 
-import {FilterableObjectVersionsTable} from '../ObjectsPage/ObjectVersionsPage';
+import {FilterableObjectVersionsTable} from '../ObjectsPage/ObjectVersionsTable';
 export const LLMJudgesTab: React.FC<{
   entity: string;
   project: string;
